New option: -x to generate Windows commands instead shell ones.
The -x flags is an addition to -s and -c to send to stdout the
relevant environment variables set commands.

It uses setx, available on the base system since Windows Vista and
in the ressource kit since Windows 2000.

http://technet.microsoft.com/en-us/library/cc755104.aspx

It allows to set the environment variables directly on Windows
user environment variables, allowing to have them available more
easily in GUI softwares too.